+++ composeApp/src/commonMain/kotlin/karel/hudera/spacetrace/domain/IRepository.kt
@@ -7,9 +7,9 @@ import kotlinx.coroutines.flow.Flow
 interface IRepository {
     suspend fun getPicture(): Picture
     suspend fun getArticles(): List<Article>
-    suspend fun getArticle(articleId: Int): Article
+    suspend fun getArticle(articleId: Long): Article
     suspend fun getFavoriteArticles(): Flow<List<Article>>
-    suspend fun isArticleFavorite(articleId: Int): Boolean
+    suspend fun isArticleFavorite(articleId: Long): Boolean
     suspend fun addArticleToFavorites(article: Article)
-    suspend fun removeArticleFromFavorites(articleId: Int)
+    suspend fun removeArticleFromFavorites(articleId: Long)
 }
